+++ commitlog_test.go
@@ -61,7 +61,7 @@ func TestAppendMessageSet(t *testing.T) {
 	defer l.Close()
 	defer cleanup()
 
-	set, _, err := newMessageSetFromProto(0, 0, msgs, false)
+	set, _, err := newMessageSetFromProto(0, 0, msgs)
 	require.NoError(t, err)
 
 	offsets, err := l.AppendMessageSet(set)